Relocate starter code files in backend
