+++ backend/starter_code_reference/services/interfaces/IEntityService.ts
