+++ backend/starter_code_reference/services/interfaces/simpleEntityService.ts
