+++ backend/typescript/package.json
@@ -63,9 +63,9 @@
     "@types/pg": "^7.14.10",
     "@types/umzug": "^2.3.0",
     "@types/validator": "^13.1.3",
-    "@typescript-eslint/eslint-plugin": "^6.2.0",
-    "@typescript-eslint/parser": "^6.2.0",
-    "eslint": "^8.50.0",
+    "@typescript-eslint/eslint-plugin": "5.44.0",
+    "@typescript-eslint/parser": "5.44.0",
+    "eslint": "8.44.0",
     "eslint-config-airbnb-typescript": "^12.3.1",
     "eslint-config-prettier": "^8.0.0",
     "eslint-plugin-import": "^2.22.1",
@@ -76,7 +76,7 @@
     "prettier": "^2.2.1",
     "sequelize-cli": "^6.6.3",
     "ts-jest": "^29.1.0",
-    "typescript": "^5.2.0"
+    "typescript": "^5.9.2"
   },
   "resolutions": {
     "@types/express": "^4.17.21",
