+++ frontend/src/pages/Default.tsx
@@ -6,7 +6,7 @@ import SampleContext from "../contexts/SampleContext";
 import Logout from "../components/auth/Logout";
 import RefreshCredentials from "../components/auth/RefreshCredentials";
 import DefaultResetPassword from "../components/auth/DefaultResetPassword";
-/* Note: starter code page, only used for development */
+/* Note: starter code file, active for development purposes */
 
 type ButtonProps = { text: string; path: string };
